+++ cupydo/interfaces/Cupydo.py
@@ -124,9 +124,9 @@ class CUPyDO(object):
             elif p['fluidSolver'] == 'Pfem':
                 from . import Pfem as fItf
                 fluidSolver = fItf.Pfem(p['cfdFile'], args.n, args.nogui, p['dt'])
-            elif p['fluidSolver'] == 'Flow':
-                from . import Flow as fItf
-                fluidSolver = fItf.Flow(p['cfdFile'], args.n)
+            elif p['fluidSolver'] == 'DART':
+                from . import Dart as fItf
+                fluidSolver = fItf.Dart(p['cfdFile'], args.n)
             elif p['fluidSolver'] == 'VLM':
                 from . import VLM as fItf
                 fluidSolver = fItf.VLMSolver(p['cfdFile'])
@@ -182,7 +182,7 @@ class CUPyDO(object):
 # Sample parameters list
 
 # Solvers
-# - p['fluidSolver'], fluid solvers available: SU2, Pfem, Flow, VLM
+# - p['fluidSolver'], fluid solvers available: SU2, Pfem, DART, VLM
 # - p['solidSolver'], solid solvers available: Metafor, RBMI, Modal, GetDP, SU2
 # Configuration files
 # - p['cfdFile'], path to fluid cfg file
