+++ tests/CMakeLists.txt
@@ -1,9 +1,9 @@
 # Add tests
 IF(NOT WIN32) # temporarily disabled on Windows
-    MACRO_AddTest(${CMAKE_CURRENT_SOURCE_DIR}/Flow_RBM)
+    MACRO_AddTest(${CMAKE_CURRENT_SOURCE_DIR}/DART_RBM)
 ENDIF()
-MACRO_AddTest(${CMAKE_CURRENT_SOURCE_DIR}/Flow_Modal)
-MACRO_AddTest(${CMAKE_CURRENT_SOURCE_DIR}/Flow_Metafor)
+MACRO_AddTest(${CMAKE_CURRENT_SOURCE_DIR}/DART_Modal)
+MACRO_AddTest(${CMAKE_CURRENT_SOURCE_DIR}/DART_Metafor)
 MACRO_AddTest(${CMAKE_CURRENT_SOURCE_DIR}/PFEM_Metafor)
 IF(NOT WIN32)
     MACRO_AddTest(${CMAKE_CURRENT_SOURCE_DIR}/SU2_Metafor)
