+++ tests/DART_Metafor/models/diamond_fluid.geo
