+++ tests/DART_Metafor/models/diamond_solid.geo
