+++ tests/DART_Metafor/staticDiamond_solid.py
