+++ tests/DART_RBM/staticNaca_solid.cfg
@@ -17,7 +17,7 @@ UNSTEADY_SIMULATION = NO
 % CSD solver (NATIVE or METAFOR)
 CSD_SOLVER = NATIVE
 %
-MESH_FILE = ../../tests/Flow_RBM/models/n0012_solid.su2
+MESH_FILE = ../../tests/DART_RBM/models/n0012_solid.su2
 %
 MOVING_MARKER = airfoil
 %
